Version du controller Etudiant sans l'execute Query

+++ src/controllers/etudiant_controller.php
@@ -43,7 +43,6 @@ class EtudiantController extends UserController{
         $niveau = $user->getNiveau();
         $ecole = $user->getEcole();
         $email = $user->getEmail();
-        echo "Le niveau : $niveau ; lecole : $ecole; lemail : $email";
         $stmt->bind_param("sss", $niveau, $ecole, $email);
     
         if ($stmt->execute()) {
@@ -109,50 +108,89 @@ class EtudiantController extends UserController{
      * @return Equipe L'objet Equipe qui représente la nouvelle équipe dans la base de données.
      * @throws Exception Si le nom de l'équipe est déjà utilisé ou si l'étudiant est déjà inscrit à un autre projet dans le même DataChallenge.
      */
-    public function registerToDataChallenge(ProjetData $projetData, Etudiant $capitaine, $nomEquipe) 
+    public function registerToDataChallenge(ProjetData $projetData, Etudiant $capitaine, $nomEquipe)
     {
         $this->conn->autocommit(false);
         try {
             // 1. Vérifier que le nom de l'équipe n'existe pas déjà
             $sql = "SELECT COUNT(*) FROM Equipes WHERE Nom = ?";
-            $count = $this->executeQuery($sql, "s", $nomEquipe);
-            if($count > 0) {
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
+            $stmt->bind_param("s", $nomEquipe);
+            $stmt->execute();
+            $stmt->bind_result($count);
+            $stmt->fetch();
+            $stmt->close();
+    
+            if ($count > 0) {
                 throw new Exception("nom equipe deja utilisé");
             }
-
+    
             // 2. Vérifier que l'étudiant n'est pas déjà inscrit à un autre ProjetData dans ce DefiData
             // Obtenir le DataChallenge du projet
             $sql = "SELECT ID_DataChallenge FROM Projets WHERE ID = ?";
-            $dataChallengeID = $this->executeQuery($sql, "i", $projetData->getId());
-
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
+            $stmt->bind_param("i", $projetData->getId());
+            $stmt->execute();
+            $stmt->bind_result($dataChallengeID);
+            $stmt->fetch();
+            $stmt->close();
+    
             // Vérifier si l'étudiant est déjà inscrit à un autre projet dans le même DataChallenge
             $sql = "SELECT COUNT(*) 
-            FROM Utilisateurs user
-            JOIN MembresEquipe ON user.ID = MembresEquipe.ID_Utilisateur
-            JOIN Equipes equipe ON MembresEquipe.ID_Equipe = equipe.ID
-            JOIN Projets projet ON equipe.ID_Projet = projet.ID
-            WHERE user.ID = ? AND projet.ID_DataChallenge = ?";
-            $count = $this->executeQuery($sql, "ii", $capitaine->getID(), $dataChallengeID);
-            if($count > 0) {
+                    FROM Utilisateurs user
+                    JOIN MembresEquipe ON user.ID = MembresEquipe.ID_Utilisateur
+                    JOIN Equipes equipe ON MembresEquipe.ID_Equipe = equipe.ID
+                    JOIN Projets projet ON equipe.ID_Projet = projet.ID
+                    WHERE user.ID = ? AND projet.ID_DataChallenge = ?";
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
+            $stmt->bind_param("ii", $capitaine->getID(), $dataChallengeID);
+            $stmt->execute();
+            $stmt->bind_result($count);
+            $stmt->fetch();
+            $stmt->close();
+    
+            if ($count > 0) {
                 throw new Exception("deja inscris");
             }
-
+    
             // L'étudiant peut s'inscrire à ce projet
             // 3. Créer une nouvelle équipe
             $equipe = new Equipe(null, $nomEquipe, [], $capitaine->getID());
-
+    
             // Insérer la nouvelle équipe dans la base de données
             $sql = "INSERT INTO Equipes (Nom, ID_Projet, ID_Capitaine) VALUES (?, ?, ?)";
-            $equipeID = $this->executeInsertQuery($sql, "sii", $equipe->getNom(), $projetData->getId(), $equipe->getChefEquipe());
-
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
+            $stmt->bind_param("sii", $equipe->getNom(), $projetData->getId(), $equipe->getChefEquipe());
+            $stmt->execute();
+            $equipeID = $stmt->insert_id;
+            $stmt->close();
+    
             // Ajouter le capitaine à cette équipe
             $sql = "INSERT INTO MembresEquipe (ID_Equipe, ID_Utilisateur) VALUES (?, ?)";
-            $this->executeInsertQuery($sql, "ii", $equipeID, $capitaine->getID());
-
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
+            $stmt->bind_param("ii", $equipeID, $capitaine->getID());
+            $stmt->execute();
+            $stmt->close();
+    
             // Mettre à jour l'ID de l'équipe et les membres de l'équipe dans l'objet Equipe
             $equipe->setId($equipeID);
             $equipe->setMembres(array($capitaine->getID()));
-
+    
             $this->conn->commit();
             return $equipe;
         } catch (Exception $e) {
@@ -162,20 +200,6 @@ class EtudiantController extends UserController{
         } finally {
             $this->conn->autocommit(true);
         }
-    }
-
-
-    private function executeQuery($sql, $types, ...$params) {
-        $stmt = $this->conn->prepare($sql);
-        if($stmt === false) {
-            throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
-        }
-        $stmt->bind_param($types, ...$params);
-        $stmt->execute();
-        $stmt->bind_result($result); //Ligne 152
-        $stmt->fetch();
-        $stmt->close();
-        return $result;
     }
     
     private function executeInsertQuery($sql, $types, ...$params) {
@@ -269,13 +293,13 @@ class EtudiantController extends UserController{
 
     /*removeMemberFromTeam() : Permet au capitaine de retirer un membre de son équipe. */
     public function removeMemberFromTeam(Equipe $equipe, Etudiant $capitaine, Etudiant $member)
-    {  
+    {
         // 1. Vérifier que le capitaine est bien le chef de l'équipe
         if ($equipe->getChefEquipe() != $capitaine->getID()) {
             throw new Exception("Le capitaine n'est pas le chef de l'équipe");
         }
     
-        if($equipe == null){
+        if ($equipe == null) {
             throw new Exception("L'équipe n'existe pas");
         }
     
@@ -291,7 +315,15 @@ class EtudiantController extends UserController{
         try {
             // 3. Retirer l'étudiant de l'équipe dans la base de données
             $sql = "DELETE FROM MembresEquipe WHERE ID_Equipe = ? AND ID_Utilisateur = ?";
-            $this->executeQuery($sql, "ii", $equipe->getId(), $member->getID());
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
+            $id_equipe = $equipe->getId();
+            $id_member = $member->getID();
+            $stmt->bind_param("ii", $id_equipe, $id_member);
+            $stmt->execute();
+            $stmt->close();
     
             // 4. Retirer l'étudiant de l'équipe dans l'objet Equipe
             $key = array_search($member->getID(), $membres);
@@ -310,8 +342,6 @@ class EtudiantController extends UserController{
         }
     }
     
-
-    /*deleteTeam() : Permet au capitaine de supprimer son équipe.*/
     public function deleteTeam(Equipe $equipe, Etudiant $capitaine)
     {
         // 1. Vérifier que le capitaine est bien le chef de l'équipe
@@ -320,7 +350,7 @@ class EtudiantController extends UserController{
             throw new Exception("Le capitaine n'est pas le chef de l'équipe");
         }
     
-        if($equipe == null){
+        if ($equipe == null) {
             throw new Exception("L'équipe n'existe pas");
         }
     
@@ -329,11 +359,23 @@ class EtudiantController extends UserController{
         try {
             // 2. Supprimer tous les membres de l'équipe dans la base de données
             $sql = "DELETE FROM MembresEquipe WHERE ID_Equipe = ?";
-            $this->executeQuery($sql, "i", $equipe->getId());
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
+            $stmt->bind_param("i", $equipe->getId());
+            $stmt->execute();
+            $stmt->close();
     
             // 3. Supprimer l'équipe dans la base de données
             $sql = "DELETE FROM Equipes WHERE ID = ?";
-            $this->executeQuery($sql, "i", $equipe->getId());
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
+            $stmt->bind_param("i", $equipe->getId());
+            $stmt->execute();
+            $stmt->close();
     
             $this->conn->commit();
             return true;
@@ -344,6 +386,7 @@ class EtudiantController extends UserController{
             $this->conn->autocommit(true);
         }
     }
+    
     
     
 
@@ -358,7 +401,7 @@ class EtudiantController extends UserController{
             throw new Exception("Le questionnaire n'est pas ouvert");
         }
     
-        if($equipe == null){
+        if ($equipe == null) {
             throw new Exception("L'équipe n'existe pas");
         }
     
@@ -368,15 +411,21 @@ class EtudiantController extends UserController{
         try {
             // 2. Pour chaque réponse, insérer une ligne dans la table Reponses
             $sql = "INSERT INTO Reponses (Contenu, Note, ID_Question, ID_Equipe) VALUES (?, ?, ?, ?)";
+            $stmt = $this->conn->prepare($sql);
+            if ($stmt === false) {
+                throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+            }
             
             foreach ($reponses as $reponse) {
                 $contenu = $reponse->getContenu();
                 $note = $reponse->getNote(); // Note peut être null si la réponse n'a pas encore été notée
                 $idQuestion = $reponse->getIdQuestion();
-    
-                $this->executeQuery($sql, "siii", $contenu, $note, $idQuestion, $equipe->getId());
+                
+                $stmt->bind_param("siii", $contenu, $note, $idQuestion, $equipe->getId());
+                $stmt->execute();
             }
-    
+            
+            $stmt->close();
             $this->conn->commit();
             return true;
         } catch (Exception $e) {
@@ -390,15 +439,21 @@ class EtudiantController extends UserController{
     
     
     /*submitCode() : Permet à l'étudiant de soumettre le lien de l'hébergement de son code (par exemple, un lien vers un référentiel GitLab). */
-    public function submitCode(ProjetData $projet, $link) 
+    public function submitCode(ProjetData $projet, $link)
     {
         // 1. Insérer le lien dans la table Ressources en tant que ressource 'Notebook' pour le projet de l'équipe
         $sql = "INSERT INTO Ressources (URL, Type, ID_Projet) VALUES (?, 'Notebook', ?)";
-    
-        $this->executeQuery($sql, "si", $link, $projet->getId());
+        $stmt = $this->conn->prepare($sql);
+        if ($stmt === false) {
+            throw new Exception('prepare() failed: ' . htmlspecialchars($this->conn->error));
+        }
+        $stmt->bind_param("si", $link, $projet->getId());
+        $stmt->execute();
+        $stmt->close();
     
         return true;
     }
+    
     
     
     /*viewDataChallenges() : Permet à l'étudiant de voir une liste de tous les défis de données disponibles. */
